refactor(oracle)!: remove `oracle get`
BREAKING CHANGE: `aecli oracle get` removed
Use `aecli inspect` instead.
